+++ include/linux/pci.h
@@ -1582,7 +1582,7 @@ enum pci_fixup_pass {
 				  class_shift, hook)			\
 	static const struct pci_fixup __PASTE(__pci_fixup_##name,__LINE__) __used	\
 	__attribute__((__section__(#section), aligned((sizeof(void *)))))    \
-		= { vendor, device, class, class_shift, hook };
+		= { (u16)vendor, (u16)device, class, class_shift, hook };
 
 #define DECLARE_PCI_FIXUP_CLASS_EARLY(vendor, device, class,		\
 					 class_shift, hook)		\
