+++ scripts/mod/file2alias.c
@@ -156,7 +156,7 @@ static void device_id_check(const char *modname, const char *device_id,
 			    unsigned long size, unsigned long id_size,
 			    void *symval)
 {
-	int i;
+	unsigned int i;
 
 	if (size % id_size || size < id_size) {
 		fatal("%s: sizeof(struct %s_device_id)=%lu is not a modulo "
@@ -185,7 +185,7 @@ static void device_id_check(const char *modname, const char *device_id,
 /* USB is special because the bcdDevice can be matched against a numeric range */
 /* Looks like "usb:vNpNdNdcNdscNdpNicNiscNipNinN" */
 static void do_usb_entry(void *symval,
-			 unsigned int bcdDevice_initial, int bcdDevice_initial_digits,
+			 unsigned int bcdDevice_initial, unsigned int bcdDevice_initial_digits,
 			 unsigned char range_lo, unsigned char range_hi,
 			 unsigned char max, struct module *mod)
 {
@@ -295,7 +295,7 @@ static void do_usb_entry_multi(void *symval, struct module *mod)
 {
 	unsigned int devlo, devhi;
 	unsigned char chi, clo, max;
-	int ndigits;
+	unsigned int ndigits;
 
 	DEF_FIELD(symval, usb_device_id, match_flags);
 	DEF_FIELD(symval, usb_device_id, idVendor);
@@ -383,8 +383,8 @@ static int do_hid_entry(const char *filename,
 	sprintf(alias, "hid:");
 	ADD(alias, "b", bus != HID_BUS_ANY, bus);
 	ADD(alias, "g", group != HID_GROUP_ANY, group);
-	ADD(alias, "v", vendor != HID_ANY_ID, vendor);
-	ADD(alias, "p", product != HID_ANY_ID, product);
+	ADD(alias, "v", vendor != (__u32)HID_ANY_ID, vendor);
+	ADD(alias, "p", product != (__u32)HID_ANY_ID, product);
 
 	return 1;
 }
@@ -431,10 +431,10 @@ static int do_pci_entry(const char *filename,
 	DEF_FIELD(symval, pci_device_id, class_mask);
 
 	strcpy(alias, "pci:");
-	ADD(alias, "v", vendor != PCI_ANY_ID, vendor);
-	ADD(alias, "d", device != PCI_ANY_ID, device);
-	ADD(alias, "sv", subvendor != PCI_ANY_ID, subvendor);
-	ADD(alias, "sd", subdevice != PCI_ANY_ID, subdevice);
+	ADD(alias, "v", vendor != (__u32)PCI_ANY_ID, vendor);
+	ADD(alias, "d", device != (__u32)PCI_ANY_ID, device);
+	ADD(alias, "sv", subvendor != (__u32)PCI_ANY_ID, subvendor);
+	ADD(alias, "sd", subdevice != (__u32)PCI_ANY_ID, subdevice);
 
 	baseclass = (class) >> 16;
 	baseclass_mask = (class_mask) >> 16;
@@ -576,7 +576,7 @@ static void do_pnp_device_entry(void *symval, unsigned long size,
 	for (i = 0; i < count; i++) {
 		DEF_FIELD_ADDR(symval + i*id_size, pnp_device_id, id);
 		char acpi_id[sizeof(*id)];
-		int j;
+		unsigned int j;
 
 		buf_printf(&mod->dev_table_buf,
 			   "MODULE_ALIAS(\"pnp:d%s*\");\n", *id);
@@ -605,7 +605,7 @@ static void do_pnp_card_entries(void *symval, unsigned long size,
 
 		for (j = 0; j < PNP_MAX_DEVICES; j++) {
 			const char *id = (char *)(*devs)[j].id;
-			int i2, j2;
+			unsigned int i2, j2;
 			int dup = 0;
 
 			if (!id[0])
@@ -631,7 +631,7 @@ static void do_pnp_card_entries(void *symval, unsigned long size,
 			/* add an individual alias for every device entry */
 			if (!dup) {
 				char acpi_id[PNP_ID_LEN];
-				int k;
+				unsigned int k;
 
 				buf_printf(&mod->dev_table_buf,
 					   "MODULE_ALIAS(\"pnp:d%s*\");\n", id);
@@ -912,7 +912,7 @@ ADD_TO_DEVTABLE("virtio", virtio_device_id, do_virtio_entry);
 static int do_vmbus_entry(const char *filename, void *symval,
 			  char *alias)
 {
-	int i;
+	unsigned int i;
 	DEF_FIELD_ADDR(symval, hv_vmbus_device_id, guid);
 	char guid_name[(sizeof(*guid) + 1) * 2];
 
@@ -981,7 +981,7 @@ static void dmi_ascii_filter(char *d, const char *s)
 static int do_dmi_entry(const char *filename, void *symval,
 			char *alias)
 {
-	int i, j;
+	unsigned int i, j;
 	DEF_FIELD_ADDR(symval, dmi_system_id, matches);
 	sprintf(alias, "dmi*");
 
@@ -1073,8 +1073,8 @@ static int do_ipack_entry(const char *filename,
 	DEF_FIELD(symval, ipack_device_id, device);
 	strcpy(alias, "ipack:");
 	ADD(alias, "f", format != IPACK_ANY_FORMAT, format);
-	ADD(alias, "v", vendor != IPACK_ANY_ID, vendor);
-	ADD(alias, "d", device != IPACK_ANY_ID, device);
+	ADD(alias, "v", vendor != (__u32)IPACK_ANY_ID, vendor);
+	ADD(alias, "d", device != (__u32)IPACK_ANY_ID, device);
 	add_wildcard(alias);
 	return 1;
 }
